fix(channex): act on the request-side audit against the vendored docs
Every Channex call was checked against docs/vendor/channex by one agent per
area and a refuter per finding (docs/audits/2026-09-17-channex-request-audit.md);
the open questions were settled by probes on staging before changing anything:

- a duplicate rate plan title on a property is refused (422 "Duplication in
  Rate Plan title is not allowed!")
- `days` on POST /availability is honoured although undocumented (kept)
- a past date_from on /restrictions is moved to today with a string warning
- values=[good, bad] answers one warning echoing the bad entry, no index
- test_connection with an unknown hotel answers 200 {success:false, errors:null}
- GET /rate_plans carries no parent_rate_plan_id; /rate_plans/options does
- GET /reviews accepts filter[updated_at][gte]

Provisioning: titles scoped by room type on Channex when a property has several
room types (adoption strips the suffix), per-room option at the room's maximum
occupancy, derived plans created with every inherit_* flag off, parent read from
/rate_plans/options. ARI: warnings matched to the sent entry by echoed keys, past
cells failed locally, read-back rates parsed in each plan's currency, per-occupancy
rates excluded from verify. Bookings: per-room taxes, collected (withheld) taxes
and services; virtual card balance and window; unique_id never a revision id;
Booking-by-id parsed by its own shape; GET /booking_revisions/:id; the unacked
feed re-read from page 1 after acks. Channels: readiness from
POST /channels/{id}/check_readiness, test_connection reads success, mapping edits
sent with PUT /channels/{id}, removal deactivates then deletes, `status` read only
where documented. Messaging and reviews: attachments as relative links, reviews
paged and filtered on updated_at with reply_error → failed response,
updated_review triggers a sync, special offers are whole numbers. Webhooks:
disconnect_channel/disconnect_listing (P1), sync_warning and rate_error recorded;
an oversized delivery kept as its event name and answered 200.

Fixtures corrected (invented parent_rate_plan_id, status, readiness_issues
removed) and added for every new call; contract, unit and integration tests
cover each change. Spec 05 and 09, CHANGELOG.

+++ apps/web/src/app/(console)/reservations/requests/requests.actions.ts
@@ -24,7 +24,11 @@ export const listRequests = withPermission<
 
 const decisionSchema = z.discriminatedUnion("decision", [
   z.object({ decision: z.literal("preapproval"), blockInstantBooking: z.boolean() }),
-  z.object({ decision: z.literal("special_offer"), totalPrice: z.coerce.number().positive() }),
+  // docs: Airbnb API › Special offer — a whole number in the listing's currency
+  z.object({
+    decision: z.literal("special_offer"),
+    totalPrice: z.coerce.number().int().positive(),
+  }),
   z.object({ decision: z.literal("accept") }),
   z.object({
     decision: z.literal("decline"),
